+++ src/ambition/version_details.cc
@@ -35,6 +35,7 @@ std::vector<VersionUpdates> getVersionDetails(
         .heading = HEADING_FIXED,
         .items = {
           _("Fixed the Inn guest selection sometimes changing when a guest leaves, resulting in sometimes hiring the wrong unit."),
+          _("Fixed the Inn hire button not getting disabled properly when a guest leaves, resulting in sometimes hiring the wrong unit."),
           _("Allow `shift+=` to be bound using the key sequence `+`."),
           _("Give users of Programmer Dvorak access to their `$`, `!`, and `#` keys.  These can be bound to actions with and without modifiers."),
         },