+++ CHANGELOG.md
@@ -34,6 +34,8 @@ speaking, considered to be the player's experience.
 
 - Fixed the Inn guest selection sometimes changing when a guest leaves,
   resulting in sometimes hiring the wrong unit.
+- Fixed the Inn hire button not getting disabled properly when a guest leaves,
+  resulting in sometimes hiring the wrong unit.
 - Allow `shift+=` to be bound using the key sequence `+`.
 - Give users of Programmer Dvorak access to their `$`, `!`, and `#` keys.  These
   can be bound to actions with and without modifiers.